// src/interfaces/signup-interface.ts
export interface SignupDetails {
    password: string
    day: string
    month: string
    year: string
    firstName: string
    lastName: string
    company: string
    address: string
    address2: string
    country: string
    state: string
    city: string
    zipcode: string
    mobileNumber: string
}